BackEnd/source/requirements.txt
django==3.0.7
djangorestframework==3.11.0
psycopg2-binary==2.8.4
django-filter==2.2.0
# ipdb
django-extensions==2.2.8
Werkzeug==1.0.0
gunicorn==19.9.0 
# django-allauth
# djangorestframework-jwt
django-cors-headers
# psycopg2==2.8.4